use WindowSize directly
