+++ core/testing/src/main/java/io/github/droidkaigi/confsched/testing/robot/FavoritesScreenRobot.kt
@@ -1,9 +1,5 @@
 package io.github.droidkaigi.confsched.testing.robot
 
-import androidx.compose.material3.windowsizeclass.ExperimentalMaterial3WindowSizeClassApi
-import androidx.compose.material3.windowsizeclass.WindowWidthSizeClass
-import androidx.compose.material3.windowsizeclass.calculateWindowSizeClass
-import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.hasTestTag
 import androidx.compose.ui.test.hasText
@@ -18,7 +14,6 @@ import io.github.droidkaigi.confsched.favorites.section.FavoritesScreenEmptyView
 import io.github.droidkaigi.confsched.model.TimetableItemId
 import io.github.droidkaigi.confsched.ui.component.TimetableItemCardBookmarkButtonTestTag
 import io.github.droidkaigi.confsched.ui.component.TimetableItemCardTestTag
-import io.github.droidkaigi.confsched.ui.compositionlocal.LocalIsWideWidthScreen
 import javax.inject.Inject
 
 class FavoritesScreenRobot @Inject constructor(
@@ -30,22 +25,12 @@ class FavoritesScreenRobot @Inject constructor(
     TimetableServerRobot by timetableServerRobot,
     DeviceQualifierRobot by deviceQualifierRobot {
 
-    @OptIn(ExperimentalMaterial3WindowSizeClassApi::class)
     fun setupFavoritesScreenContent() {
         robotTestRule.setContent {
-            val windowSize = calculateWindowSizeClass()
-            val isWideWidthScreen = when (windowSize.widthSizeClass) {
-                WindowWidthSizeClass.Compact -> false
-                WindowWidthSizeClass.Medium -> true
-                WindowWidthSizeClass.Expanded -> true
-                else -> false
-            }
-            CompositionLocalProvider(LocalIsWideWidthScreen provides isWideWidthScreen) {
-                KaigiTheme {
-                    FavoritesScreen(
-                        onTimetableItemClick = {},
-                    )
-                }
+            KaigiTheme {
+                FavoritesScreen(
+                    onTimetableItemClick = {},
+                )
             }
         }
         waitUntilIdle()
