+++ core/testing/src/main/java/io/github/droidkaigi/confsched/testing/robot/SearchScreenRobot.kt
@@ -1,9 +1,5 @@
 package io.github.droidkaigi.confsched.testing.robot
 
-import androidx.compose.material3.windowsizeclass.ExperimentalMaterial3WindowSizeClassApi
-import androidx.compose.material3.windowsizeclass.WindowWidthSizeClass
-import androidx.compose.material3.windowsizeclass.calculateWindowSizeClass
-import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.ui.test.assertAll
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.assertIsNotDisplayed
@@ -33,7 +29,6 @@ import io.github.droidkaigi.confsched.testing.utils.hasTestTag
 import io.github.droidkaigi.confsched.ui.Inject
 import io.github.droidkaigi.confsched.ui.component.TimetableItemCardTestTag
 import io.github.droidkaigi.confsched.ui.component.TimetableItemCardTitleTextTestTag
-import io.github.droidkaigi.confsched.ui.compositionlocal.LocalIsWideWidthScreen
 
 const val DemoSearchWord = "Demo"
 
@@ -68,23 +63,13 @@ class SearchScreenRobot @Inject constructor(
         ENGLISH("EN"),
     }
 
-    @OptIn(ExperimentalMaterial3WindowSizeClassApi::class)
     fun setupSearchScreenContent() {
         robotTestRule.setContent {
-            val windowSize = calculateWindowSizeClass()
-            val isWideWidthScreen = when (windowSize.widthSizeClass) {
-                WindowWidthSizeClass.Compact -> false
-                WindowWidthSizeClass.Medium -> true
-                WindowWidthSizeClass.Expanded -> true
-                else -> false
-            }
-            CompositionLocalProvider(LocalIsWideWidthScreen provides isWideWidthScreen) {
-                KaigiTheme {
-                    SearchScreen(
-                        onTimetableItemClick = {},
-                        onBackClick = {},
-                    )
-                }
+            KaigiTheme {
+                SearchScreen(
+                    onTimetableItemClick = {},
+                    onBackClick = {},
+                )
             }
         }
         waitUntilIdle()
