--- core/ui/src/commonMain/kotlin/io/github/droidkaigi/confsched/ui/compositionlocal/LocalIsWideWidthScreen.kt
@@ -1,6 +1,0 @@
-package io.github.droidkaigi.confsched.ui.compositionlocal
-
-import androidx.compose.runtime.compositionLocalOf
-
-@Suppress("CompositionLocalAllowlist")
-val LocalIsWideWidthScreen = compositionLocalOf { false }
