+++ feature/sessions/src/commonMain/kotlin/io/github/droidkaigi/confsched/sessions/section/TimetableList.kt
@@ -17,6 +17,9 @@ import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.LazyListState
 import androidx.compose.foundation.lazy.items
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.windowsizeclass.ExperimentalMaterial3WindowSizeClassApi
+import androidx.compose.material3.windowsizeclass.WindowWidthSizeClass
+import androidx.compose.material3.windowsizeclass.calculateWindowSizeClass
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
@@ -33,7 +36,6 @@ import io.github.droidkaigi.confsched.sessions.timetableDetailSharedContentState
 import io.github.droidkaigi.confsched.ui.component.TimetableItemCard
 import io.github.droidkaigi.confsched.ui.component.TimetableItemTag
 import io.github.droidkaigi.confsched.ui.compositionlocal.LocalAnimatedVisibilityScope
-import io.github.droidkaigi.confsched.ui.compositionlocal.LocalIsWideWidthScreen
 import io.github.droidkaigi.confsched.ui.compositionlocal.LocalSharedTransitionScope
 import io.github.droidkaigi.confsched.ui.icon
 import kotlinx.collections.immutable.PersistentMap
@@ -52,7 +54,7 @@ data class TimetableListUiState(
     }
 }
 
-@OptIn(ExperimentalSharedTransitionApi::class)
+@OptIn(ExperimentalSharedTransitionApi::class, ExperimentalMaterial3WindowSizeClassApi::class)
 @Composable
 fun TimetableList(
     uiState: TimetableListUiState,
@@ -66,7 +68,13 @@ fun TimetableList(
     val layoutDirection = LocalLayoutDirection.current
     val sharedTransitionScope = LocalSharedTransitionScope.current
     val animatedScope = LocalAnimatedVisibilityScope.current
-    val isWideWidthScreen = LocalIsWideWidthScreen.current
+    val windowSize = calculateWindowSizeClass()
+    val isWideWidthScreen = when (windowSize.widthSizeClass) {
+        WindowWidthSizeClass.Compact -> false
+        WindowWidthSizeClass.Medium -> true
+        WindowWidthSizeClass.Expanded -> true
+        else -> false
+    }
     val columnNum by remember { derivedStateOf { if (isWideWidthScreen) 2 else 1 } }
 
     LazyColumn(
